Create attribute interface
The attribute interface has been created with the following
functions:
  * `inquery_attrbute_new()`
  * `inquery_attribute_new_nil()`
  * `inquery_attribute_new_int()`
  * `inquery_attribute_new_real()`
  * `inquery_attribute_new_text()`
  * `inquery_attribute_copy()`
  * `inquery_attribute_free()`
  * `inquery_attribute_key()`
  * `inquery_attribute_value()`
  * `inquery_attribute_type()`
  * `inquery_attribute_cmp()`
  * `inquery_attribute_lt()`
  * `inquery_attribute_lteq()`
  * `inquery_attribute_eq()`
  * `inquery_attribute_gteq()`
  * `inquery_attribute_gt()`
  * `inquery_attribute_nil()`
  * `inquery_attribute_int()`
  * `inquery_attribute_real()`
  * `inquery_attribute_text()`
  * `inquery_attribute_string()`
  * `inquery_attribute_json()`

Supporting this interface is the `inquery_attribute` type,
which is  specialisation of the `inquery_object` type. The
`inquery_attribute_smart` type has also been defined as a
specialised form of `inquery_attribute` that supports
automatic release when it is out of scope.

Parallel to the `inquery_attribute` interface, the
`inquery_value` interface has also been created to handle
the functionality related to boxed values which are contained
within attributes. Although the interface for `inquery_value`
is currently public, I plan to merge it soon into the
attribute interface.

This commit closes issue #14.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>

Add tests for attribute interface

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>